requirements.txt
Brotli==1.0.9
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
GeoAlchemy2==0.14.4
greenlet==3.0.1
idna==3.6
logging
numexpr==2.8.7
numpy==1.26.3
pandas==2.1.4
pip==23.3.1
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wheel==0.41.2